WIP: Fix occlusion query not working after a clear with draw
Metal doesn't support visiblity buffer's offset reuse within a render pass
so allocate multiple offsets for one query in this case, if it is interrupted
by a clear with draw command.

+++ src/libANGLE/renderer/metal/mtl_common.h
@@ -142,7 +142,7 @@ constexpr MTLVertexStepFunction kVertexStepFunctionInvalid =
 
 constexpr float kEmulatedAlphaValue = 1.0f;
 
-constexpr size_t kOcclusionQueryResultSize = sizeof(uint64_t);
+constexpr uint32_t kOcclusionQueryResultSize = sizeof(uint64_t);
 
 // NOTE(hqle): Support ES 3.0.
 constexpr gl::Version kMaxSupportedGLVersion = gl::Version(2, 0);
